+++ examples/log_analytics/entity/log_analytics_entity_associations_remove.tf
@@ -1,0 +1,14 @@
+// Copyright (c) 2017, 2025, Oracle and/or its affiliates. All rights reserved.
+// Licensed under the Mozilla Public License v2.0
+
+variable "log_analytics_entity_associations_remove_association_entities" {
+  default = []
+}
+
+# Remove entity associations with required parameters
+resource "oci_log_analytics_log_analytics_entity_associations_remove" "test_log_analytics_entity_associations_remove" {
+  #Required
+  association_entities    = tolist([oci_log_analytics_log_analytics_entity.entityOptional.id])
+  log_analytics_entity_id = oci_log_analytics_log_analytics_entity.entityRequired.id
+  namespace               = data.oci_objectstorage_namespace.ns.namespace
+}