+++ internal/service/generative_ai/generative_ai_imported_model_data_source.go
@@ -1,0 +1,71 @@
+// Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
+// Licensed under the Mozilla Public License v2.0
+
+package generative_ai
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	oci_generative_ai "github.com/oracle/oci-go-sdk/v65/generativeai"
+
+	"github.com/oracle/terraform-provider-oci/internal/client"
+	"github.com/oracle/terraform-provider-oci/internal/tfresource"
+)
+
+func GenerativeAiImportedModelDataSource() *schema.Resource {
+	fieldMap := make(map[string]*schema.Schema)
+	fieldMap["imported_model_id"] = &schema.Schema{
+		Type:     schema.TypeString,
+		Required: true,
+	}
+	return tfresource.GetSingularDataSourceItemSchemaWithContext(GenerativeAiImportedModelResource(), fieldMap, readSingularGenerativeAiImportedModelWithContext)
+}
+
+func readSingularGenerativeAiImportedModelWithContext(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+	sync := &GenerativeAiImportedModelDataSourceCrud{}
+	sync.D = d
+	sync.Client = m.(*client.OracleClients).GenerativeAiClient()
+
+	return tfresource.HandleDiagError(m, tfresource.ReadResourceWithContext(ctx, sync))
+}
+
+type GenerativeAiImportedModelDataSourceCrud struct {
+	D      *schema.ResourceData
+	Client *oci_generative_ai.GenerativeAiClient
+	Res    *oci_generative_ai.GetImportedModelResponse
+}
+
+func (s *GenerativeAiImportedModelDataSourceCrud) VoidState() {
+	s.D.SetId("")
+}
+
+func (s *GenerativeAiImportedModelDataSourceCrud) GetWithContext(ctx context.Context) error {
+	request := oci_generative_ai.GetImportedModelRequest{}
+
+	if importedModelId, ok := s.D.GetOkExists("imported_model_id"); ok {
+		tmp := importedModelId.(string)
+		request.ImportedModelId = &tmp
+	}
+
+	request.RequestMetadata.RetryPolicy = tfresource.GetRetryPolicy(false, "generative_ai")
+
+	response, err := s.Client.GetImportedModel(ctx, request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response
+	return nil
+}
+
+func (s *GenerativeAiImportedModelDataSourceCrud) SetData() error {
+	if s.Res == nil {
+		return nil
+	}
+
+	s.D.SetId(*s.Res.Id)
+
+	return nil
+}